ALL: Revert r1037:1041. Fix build.
FEX is not a common oss library thus it is not packaged by open source operating system distributors. That means if we ever want VBA-M to be packaged in OS distributions, we need to embed it in our main source tree, since fex is a dependency. 
The "dependencies" folder is outside of our main source tree thus it cant be used when building from a source tarball (for packaging).
zlib and SFML from the "dependencies" folder are only used when building the win32 port. All the other ports only use system libraries.
Having FEX somewhere within the "trunk" folder is REQUIRED for packaging, moving it elsewhere is NOT AN OPTION. Please live with it.


git-svn-id: https://svn.code.sf.net/p/vbam/code/trunk@1042 a31d4220-a93d-0410-bf67-fe4944624d44