Added E-Reader scan support by CaitSith2.
One thing it does, is correct any raw files scanned, if they are correctable. (usually, uncorrectable raw files come up with a read error.)  There is also bin support included as well. (Note, if you open a valid dotcode bin, it will encode it to raw, and save the freshly encoded raw file if it has been set in the menu. If the dotcode save file has not been set, the bin will still load.)

Currently, part of the GBA emulation breaks the E-Reader scan code. The reader always returns a Region Error.