Change library name to VBA-M Plus
Due to the major changes in how battery save file is handled, the
core has to have a unique name. A BACKUP OF EXISTING SAVES IS
STILL REQUIRED!

This PR should minimize conflicts with existing vbam core. Users
that configured retroarch to use "Sort" for saves and states will have
a separate folder for the core so their existing saves and states will
not be overwritten.

Users that do not have this configured would still have the risk of
saves overwritten, although this would mostly affect savefiles that uses
EEPROM. Flash and SRAM types should probably work fine.